投稿削除に制限

+++ app/controllers/posts_controller.rb
@@ -29,8 +29,12 @@ class PostsController < ApplicationController
 
   def destroy
     post = Post.find(params[:id])
-    post.destroy
-    redirect_to request.referer
+    if post.user_id == current_user.id 
+      post.destroy
+      redirect_to request.referer
+    else
+      redirect_to root_path
+    end
   end
 
   private
